functionalize code
