+++ index.js
@@ -1,250 +1,13 @@
 require("./typedefs");
 const core = require("@actions/core");
 const github = require("@actions/github");
-const path = require("path");
-const fs = require("fs").promises;
 const checks = require("./checks");
-
-// These are all of the files that, if changed, will trigger the check suite
-const filesToCheck = ["orders", "secrets.json", "policy.json"];
-
-// We want to track how all the checks go
-const counts = {
-  success: 0,
-  failure: 0,
-  warning: 0,
-  notice: 0,
-};
-
-// Emojis are fun
-const icons = {
-  failure: "💀",
-  warning: "⚠️",
-  notice: "👉",
-};
-
-/**
- * Takes a filename like secrets.json and returns secretsJson
- * @param {string} filename
- */
-function _camelCaseFileName(filename) {
-  const words = filename.split(".");
-
-  let result = words[0];
-
-  if (words.length > 1) {
-    result += words
-      .slice(1)
-      .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-      .join();
-  }
-
-  return result;
-}
-
-/**
- * Read orders, secrets.json, and policy.json from the directory,
- * and split them by \n.
- * @param {String} filePath the path for the orders file
- * @returns {Deployment}
- */
-async function getContents(serviceName) {
-  const result = { serviceName };
-  for (let filename of filesToCheck) {
-    const filepath = path.join(serviceName, filename);
-    try {
-      await fs.stat(filepath);
-      const contents = await fs.readFile(filepath, "utf8");
-      result[`${_camelCaseFileName(filename)}Path`] = filepath;
-      result[`${_camelCaseFileName(filename)}Contents`] = contents.split("\n");
-    } catch (e) {
-      // No particular file is required in order to run the check suite
-    }
-  }
-  return result;
-}
-
-/**
- * Clear any comments from this bot that are already on the PR.
- * This prevents excessive comment polution
- * @param {Octokit} octokit
- * @param {{
- * owner: string,
- * repo: string,
- * pull_number: number
- * }} options
- */
-async function clearPreviousRunComments(octokit, { owner, repo, pull_number }) {
-  try {
-    const { data: reviewComments } = await octokit.pulls.listReviewComments({
-      owner,
-      repo,
-      pull_number,
-    });
-
-    const { data: issueComments } = await octokit.issues.listComments({
-      owner,
-      repo,
-      issue_number: pull_number,
-    });
-
-    const allDeletions = [];
-
-    reviewComments
-      .filter(
-        (c) => c.user.login === "github-actions[bot]" && c.user.type === "Bot"
-      )
-      .forEach((comment) => {
-        allDeletions.push(
-          octokit.pulls.deleteReviewComment({
-            owner,
-            repo,
-            comment_id: comment.id,
-          })
-        );
-      });
-
-    issueComments
-      .filter(
-        (c) => c.user.login === "github-actions[bot]" && c.user.type === "Bot"
-      )
-      .forEach((comment) => {
-        allDeletions.push(
-          octokit.issues.deleteComment({
-            owner,
-            repo,
-            comment_id: comment.id,
-          })
-        );
-      });
-
-    await Promise.all(allDeletions);
-  } catch (e) {
-    console.log(e);
-    throw e;
-  }
-}
-
-async function suggestBugReport(
-  octokit,
-  error,
-  title,
-  { owner, repo, pull_number: issue_number }
-) {
-  const errorText = `\`\`\`\n${error.message}\n\n${error.stack}\n\`\`\``;
-  const issueLink = `[Create an issue](https://github.com/glg-public/gds-cc-screamer/issues/new?title=${encodeURIComponent(
-    title
-  )}&body=${encodeURIComponent(errorText)})`;
-  const body = `## An error was encountered. Please submit a bug report
-${errorText}
-
-${issueLink}
-  `;
-  await octokit.issues.createComment({
-    owner,
-    repo,
-    issue_number,
-    body,
-  });
-}
-
-/**
- * Leaves the correct type of comment for a given result and deployment
- * @param {Octokit} octokit A configured octokit client
- * @param {Deployment} deployment
- * @param {Result} result
- * @param {{
- * owner: string,
- * repo: string,
- * pull_number: number,
- * sha: string
- * }} options
- */
-async function leaveComment(
-  octokit,
-  deployment,
-  result,
-  { owner, repo, pull_number, sha }
-) {
-  const { ordersPath, secretsPath, policyPath } = deployment;
-  // Build a markdown comment to post
-  let comment = `## ${icons[result.level]} ${result.title}\n`;
-  for (const problem of result.problems) {
-    comment += `- ${problem}\n`;
-    core.error(`${result.title} - ${problem}`);
-  }
-  try {
-    // Line 0 means a general comment, not a line-specific comment
-    if (result.line === 0) {
-      await octokit.issues.createComment({
-        owner,
-        repo,
-        issue_number: pull_number,
-        body: comment,
-      });
-    }
-
-    // If result.line is a range object like { start, end }, make a multi-line comment
-    else if (
-      isNaN(result.line) &&
-      result.line.hasOwnProperty("start") &&
-      result.line.hasOwnProperty("end")
-    ) {
-      await octokit.pulls.createReviewComment({
-        owner,
-        repo,
-        pull_number,
-        commit_id: sha,
-        path: result.path || ordersPath || secretsPath || policyPath,
-        body: comment,
-        side: "RIGHT",
-        start_line: result.line.start,
-        line: result.line.end,
-      });
-    }
-
-    // If line number is anything but 0, or a range object, we make a line-specific comment
-    else {
-      await octokit.pulls.createReviewComment({
-        owner,
-        repo,
-        pull_number,
-        commit_id: sha,
-        path: result.path || ordersPath || secretsPath || policyPath,
-        body: comment,
-        side: "RIGHT",
-        line: result.line,
-      });
-    }
-  } catch (e) {
-    // If the error is due to the problem existing outside the diff,
-    // we still want to alert the user, so make a generic issue comment
-    if (
-      e.errors.filter(
-        (err) =>
-          err.resource === "PullRequestReviewComment" &&
-          ["path", "line"].includes(err.field)
-      ).length > 0
-    ) {
-      result.problems.unshift(`Problem existed outside of diff at \`${result.path}\`, line **${result.line}**`);
-      result.line = 0;
-      await leaveComment(octokit, deployment, result, {
-        owner,
-        repo,
-        pull_number,
-        sha,
-      });
-    } else {
-      console.log(e);
-      console.log(result);
-      await suggestBugReport(octokit, e, "Error while posting comment", {
-        owner,
-        repo,
-        pull_number,
-      });
-    }
-  }
-}
+const {
+  clearPreviousRunComments,
+  getAllDeployments,
+  suggestBugReport,
+  leaveComment,
+} = require("./util");
 
 /**
  * Perform all checks on all deployments included in a PR
@@ -273,24 +36,23 @@ async function run() {
       pull_number,
     });
 
-    const deployments = await Promise.all(
-      Array.from(
-        new Set(
-          files
-            .filter((f) =>
-              filesToCheck.includes(path.basename(f.filename).toLowerCase())
-            )
-            .filter((f) => f.status !== "removed")
-            .map((f) => path.dirname(f.filename))
-        )
-      ).map((fn) => getContents(fn))
-    );
+    // These are all of the files that, if changed, will trigger the check suite
+    const filesToCheck = ["orders", "secrets.json", "policy.json"];
+    const deployments = await getAllDeployments(files, filesToCheck);
+
+    // We want to track how all the checks go
+    const counts = {
+      success: 0,
+      failure: 0,
+      warning: 0,
+      notice: 0,
+    };
 
     // Run every check against each deployment. Each check can have
-    // multiple results.
+    // multiple results. Each result can have multiple problems.
     for (const deployment of deployments) {
       for (const check of checks) {
-        let results;
+        let results = [];
         try {
           results = await check(deployment, github.context, inputs);
         } catch (e) {
@@ -306,6 +68,7 @@ async function run() {
         if (results.length === 0) {
           core.info("...Passed");
           counts.success += 1;
+          continue;
         }
         for (const result of results) {
           if (result.problems.length > 0) {
