--- test/root-index.js
@@ -1,5 +1,0 @@
-const { expect } = require('chai');
-
-describe('Check Running Infrastructure', () => {
-  it('Loads all present important files in a service with changes');
-});
